requirements.txt
gradio==4.19.2
python-docx==1.1.0
scikit-learn==1.4.0
huggingface-hub==0.20.3
pyyaml==6.0.1
numpy==1.26.3